creating report for each subfolder
