+++ programa/apolo_11/simulator/reports/report_dir_2024-01-16 19_51_32.771687.txt
@@ -1,0 +1,21 @@
+File 1 Summary:
+  Date: 16-01-2024 19:51:32
+  Mission: GalaxyTwo
+  Device: airplanes
+  Device Status: faulty
+  Hash: 416dadf7b1dc501a2c33473811b793dcf6f49b8d408cd9aab72618bd9cbc5c20
+
+File 2 Summary:
+  Date: 16-01-2024 19:51:32
+  Mission: GalaxyTwo
+  Device: suits
+  Device Status: warning
+  Hash: 14fa63daf3085c9a3feb245982a7ee428be0d70717d84bb85264022cc48d74f8
+
+File 3 Summary:
+  Date: 16-01-2024 19:51:37
+  Mission: ColonyMoon
+  Device: airplanes
+  Device Status: unknown
+  Hash: 7aa41967f0a7c3b5b3405f7d44983b5d2018d84b33a34cf558846986edae6410
+
